Trying a different tack for recent_activity since the old way was still producing too much data and overflowing PHP memory on busier sites (eg moodle.org).
Now, there are more database queries, which is unfortunate, but the
data is much more specific, and no sorting needs to be done, so
this is a performance boost.

I don't know how these will cancel out ... my guess is that very
small sites may be very slightly slower on the course page, but
that large sites will be much faster.

Let's see.